Agrupar los forms de usuarios en archivo form
